--- Task.Connector/Intefraces/IPermissionService.cs
@@ -1,8 +1,0 @@
-﻿using Task.Integration.Data.Models.Models;
-
-namespace Task.Connector.Intefraces;
-
-internal interface IPermissionService
-{
-	IEnumerable<Permission> GetAllPermissions();
-}
