+++ Task.Connector/Services/PermissionService/IPermissionService.cs
@@ -1,0 +1,8 @@
+﻿using Task.Integration.Data.Models.Models;
+
+namespace Task.Connector.Services.PermissionService;
+
+internal interface IPermissionService
+{
+    IEnumerable<Permission> GetAllPermissions();
+}
