+++ Task.Connector/Services/UserPermissionService/IUserPermissionService.cs
@@ -1,0 +1,8 @@
+﻿namespace Task.Connector.Services.UserPermissionService;
+
+internal interface IUserPermissionService
+{
+    IEnumerable<string> GetUserPermissions(string userLogin);
+    void AddUserPermissions(string userLogin, IEnumerable<string> rightIds);
+    void RemoveUserPermissions(string userLogin, IEnumerable<string> rightIds);
+}
